refactor(tests): split helpers out into mods
