+++ tests/integration/helpers/mod.rs
@@ -1,0 +1,2 @@
+pub mod project;
+pub mod project_builder;
